feat: add cheack commands
